3_script.rb

+++ 3_script.rb
@@ -6,6 +6,7 @@ counter = 1
 tot = Dir["data/signature_requests/*"].length
 Dir.each_child('./data/signature_requests') do |srid|
   puts "skip exist #{srid}" if File.exist?("./data/files/#{srid}")
+  next if File.exist?("./data/files/#{srid}")
   puts "Downloading #{counter}/#{tot} #{srid}"
   file_bin = client.signature_request_files :signature_request_id => srid
   open("./data/files/#{srid}", "wb") do |file|
